add new package domain for domain operation purposes
