+++ src/main/java/com/infoshare/alpha/App.java
@@ -3,13 +3,13 @@ package com.infoshare.alpha;
 import com.infoshare.alpha.wwr.common.Address;
 import com.infoshare.alpha.wwr.common.Pesel;
 import com.infoshare.alpha.wwr.di.AppDI;
-import com.infoshare.alpha.wwr.facilities.entity.Facilities;
-import com.infoshare.alpha.wwr.facilities.entity.Facility;
-import com.infoshare.alpha.wwr.facilities.query.FacilityPatientQuery;
-import com.infoshare.alpha.wwr.facilities.query.FacilityQuery;
-import com.infoshare.alpha.wwr.facilities.query.FacilityQueryField;
-import com.infoshare.alpha.wwr.facilities.readmodel.FacilitiesReadModel;
-import com.infoshare.alpha.wwr.patients.entity.Patient;
+import com.infoshare.alpha.wwr.domain.facilities.entity.Facilities;
+import com.infoshare.alpha.wwr.domain.facilities.entity.Facility;
+import com.infoshare.alpha.wwr.domain.facilities.query.FacilityPatientQuery;
+import com.infoshare.alpha.wwr.domain.facilities.query.FacilityQuery;
+import com.infoshare.alpha.wwr.domain.facilities.query.FacilityQueryField;
+import com.infoshare.alpha.wwr.domain.facilities.readmodel.FacilitiesReadModel;
+import com.infoshare.alpha.wwr.domain.patients.entity.Patient;
 
 import java.util.ArrayList;
 import java.util.HashMap;
