+++ src/main/java/com/infoshare/alpha/wwr/domain/facilities/command/FacilityDeleteCommand.java
@@ -1,0 +1,6 @@
+package com.infoshare.alpha.wwr.domain.facilities.command;
+
+public class FacilityDeleteCommand extends FacilityCommand {
+
+
+}
