+++ src/main/java/com/infoshare/alpha/wwr/domain/facilities/query/FacilityQuery.java
@@ -1,0 +1,18 @@
+package com.infoshare.alpha.wwr.domain.facilities.query;
+
+import java.util.Map;
+
+public class FacilityQuery {
+
+    private Map<FacilityQueryField, String> queryFields;
+
+
+    public FacilityQuery(Map<FacilityQueryField, String> queryFields) {
+        this.queryFields = queryFields;
+    }
+
+    public Map<FacilityQueryField, String > getQueryFields() {
+        return queryFields;
+    }
+
+}
