+++ src/main/java/com/infoshare/alpha/wwr/domain/facilities/query/FacilityQueryField.java
@@ -1,0 +1,20 @@
+package com.infoshare.alpha.wwr.domain.facilities.query;
+
+public enum FacilityQueryField {
+
+    CITY("city"),
+    STREET("street"),
+    PHONE("phone"),
+    FACILITY_NAME("facility_name");
+
+    private String allowedField;
+
+    FacilityQueryField(String allowedField) {
+        this.allowedField = allowedField;
+    }
+
+    @Override
+    public String toString() {
+        return allowedField;
+    }
+}
