+++ src/main/java/com/infoshare/alpha/wwr/domain/facilities/repository/FacilitiesDbRepository.java
@@ -1,0 +1,18 @@
+package com.infoshare.alpha.wwr.domain.facilities.repository;
+
+import com.infoshare.alpha.wwr.di.DI;
+import com.infoshare.alpha.wwr.domain.facilities.datastorage.FacilitiesJsonStorage;
+import com.infoshare.alpha.wwr.domain.facilities.entity.Facilities;
+
+public class FacilitiesDbRepository implements FacilitiesRepository, DI {
+
+	FacilitiesJsonStorage storage;
+	
+    public FacilitiesDbRepository(FacilitiesJsonStorage storage){
+        this.storage = storage;
+    }
+
+    public void persist(Facilities facilities) {
+        this.storage.save(facilities);
+    }
+}
