+++ src/main/java/com/infoshare/alpha/wwr/domain/facilities/repository/FacilitiesRepository.java
@@ -1,0 +1,9 @@
+package com.infoshare.alpha.wwr.domain.facilities.repository;
+
+import com.infoshare.alpha.wwr.domain.facilities.entity.Facilities;
+
+public interface FacilitiesRepository {
+
+    public void persist(Facilities facilities);
+
+}
