+++ src/main/java/com/infoshare/alpha/wwr/domain/patients/query/PatientFacilityQueryFields.java
@@ -1,0 +1,18 @@
+package com.infoshare.alpha.wwr.domain.patients.query;
+
+public enum PatientFacilityQueryFields {
+
+    CITY("city"),
+    FACILITY_NAME("facility_name");
+
+    private String allowedField;
+
+    PatientFacilityQueryFields(String allowedField) {
+        this.allowedField = allowedField;
+    }
+
+    @Override
+    public String toString() {
+        return allowedField;
+    }
+}
