--- src/main/java/com/infoshare/alpha/wwr/facilities/command/FacilityDeleteCommand.java
@@ -1,6 +1,0 @@
-package com.infoshare.alpha.wwr.facilities.command;
-
-public class FacilityDeleteCommand extends FacilityCommand {
-
-
-}
