--- src/main/java/com/infoshare/alpha/wwr/facilities/entity/Facilities.java
@@ -1,34 +1,0 @@
-package com.infoshare.alpha.wwr.facilities.entity;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Facilities {
-
-    private List<Facility> facilities = new ArrayList<>();
-
-    public Facilities() {
-    }
-
-    public void setFacilities(List<Facility> facilities) {
-
-        this.facilities = facilities;
-    }
-
-    public List<Facility> getFacilities() {
-
-        return facilities;
-    }
-
-    public void add(Facility facility) {
-
-        this.facilities.add(facility);
-    }
-
-    public void printAllFacilities() {
-
-        for (Facility facility : this.facilities) {
-            System.out.println(facility.toString());
-        }
-    }
-}
