--- src/main/java/com/infoshare/alpha/wwr/facilities/repository/FacilitiesRepository.java
@@ -1,9 +1,0 @@
-package com.infoshare.alpha.wwr.facilities.repository;
-
-import com.infoshare.alpha.wwr.facilities.entity.Facilities;
-
-public interface FacilitiesRepository {
-
-    public void persist(Facilities facilities);
-
-}
