--- src/main/java/com/infoshare/alpha/wwr/patients/query/PatientFacilityQueryFields.java
@@ -1,18 +1,0 @@
-package com.infoshare.alpha.wwr.patients.query;
-
-public enum PatientFacilityQueryFields {
-
-    CITY("city"),
-    FACILITY_NAME("facility_name");
-
-    private String allowedField;
-
-    PatientFacilityQueryFields(String allowedField) {
-        this.allowedField = allowedField;
-    }
-
-    @Override
-    public String toString() {
-        return allowedField;
-    }
-}
