--- src/main/java/com/infoshare/alpha/wwr/patients/readmodel/PatientsReadModelDb.java
@@ -1,16 +1,0 @@
-package com.infoshare.alpha.wwr.patients.readmodel;
-
-import com.infoshare.alpha.wwr.facilities.entity.Facility;
-import com.infoshare.alpha.wwr.patients.query.PatientFacilityQuery;
-import com.infoshare.alpha.wwr.patients.entity.Patient;
-
-import java.util.List;
-
-public interface PatientsReadModelDb {
-
-    public List<Facility> getNearestPatientFacilitiesByCity(Patient patient);
-
-    public List<Facility> getPatientFacilitiesByQuery(PatientFacilityQuery query);
-
-
-}
